Added back removed RCLCPP::spin, commented out
